feat: add pagination component to Home page and refactor project-related components
