+++ app/page.tsx
@@ -1,7 +1,16 @@
 import { ApplyStatueMenubar } from "@/components/Home/ApplyStatueMenubar";
 import { ProjectList } from "@/components/Home/ProjectList";
 import SearchCreateRow from "@/components/Home/SearchCreateRow";
-import { Pagination } from "@/components/ui/pagination";
+import {
+  Pagination,
+  PaginationContent,
+  PaginationItem,
+  PaginationLink,
+  PaginationNext,
+  PaginationPrevious,
+} from "@/components/ui/pagination";
+
+const pageNum = 5;
 
 export default async function Home() {
   return (
@@ -14,8 +23,26 @@ export default async function Home() {
           <span>• 인기순</span>
         </div>
         <ProjectList />
-        <Pagination />
       </div>
+      <Pagination className="my-8">
+        <PaginationContent>
+          <PaginationItem>
+            <PaginationPrevious href="#" />
+          </PaginationItem>
+
+          {[...Array(pageNum)].map((_, index) => {
+            return (
+              <PaginationItem key={index}>
+                <PaginationLink href="#">{index + 1}</PaginationLink>
+              </PaginationItem>
+            );
+          })}
+
+          <PaginationItem>
+            <PaginationNext href="#" />
+          </PaginationItem>
+        </PaginationContent>
+      </Pagination>
     </div>
   );
 }
