+++ components/Project/MajorGraph.tsx
